✨ Ajout d'une commande cli pour vérifier l'existance de fichiers S3 (#2292)

+++ packages/applications/cli/src/helpers/renameFile.ts
@@ -1,24 +1,24 @@
 import { DeleteObjectCommand } from '@aws-sdk/client-s3';
 
 import { getLogger } from '@potentiel-libraries/monitoring';
-
-import { copyFile } from './copyFile';
-import { assertFileExists } from './assertFileExists';
-import { getClient } from './getClient';
-import { getBucketName } from './getBucketName';
+import { fileExists, copyFile, getBucketName, getClient } from '@potentiel-libraries/file-storage';
 
 class DeleteFailedError extends Error {
   constructor() {
     super('La suppression du fichier a échoué');
   }
 }
+export class RenameFailedError extends Error {}
 
 /**
  * Renames a file in the same bucket
  * Do not use in domain, this is intended for CLI utility: files related to an immutable event are also immutable
  **/
 export const renameFile = async (fromName: string, toName: string) => {
-  await assertFileExists(fromName);
+  const exists = await fileExists(fromName);
+  if (!exists) {
+    throw new RenameFailedError(`From file not found: ${fromName}`);
+  }
 
   await copyFile(fromName, toName);
